refactor(KAN-414 F4 group 2): convert the media_uploads gate to a behavioural test
Group 2, file 1 of 8. The template for the rest.

THE INVARIANT, IN WORDS: `media_uploads` covers "Profile photo & file/media
uploads". When an admin revokes it, BOTH entrypoints must refuse —
uploadProfileFile AND uploadAvatar — and each must refuse BEFORE touching
storage. A revoke that stops only one path is a half-effective admin action,
and the half that still works is the one nobody tests.

WHAT WAS THERE: a source-text scan that regexed both files for
`getMyFeatureEntitlements` and compared string indices to guess the gate
preceded the upload. That passes if the identifier merely appears — in a
comment, in dead code, in a function nobody calls — and breaks on any file move,
which is the coupling F4 exists to end.

WHAT REPLACES IT: both actions are executed with the entitlement revoked, and
asserted to refuse with storage never reached; then executed with it GRANTED and
asserted to proceed past the gate and fail at the next check. That second
direction is the one that matters — it proves the earlier refusal came from the
gate rather than from some unrelated guard upstream.

The KAN-417 PoC only covered uploadProfileFile. Converting on that alone would
have quietly halved the coverage while looking like a modernisation, since the
block it replaces asserts BOTH entrypoints. uploadAvatar is now covered in both
directions too, and the PoC is promoted out of its `kan417-poc-` scaffold name.

MUTATION PROOF, per §8 group 2 — required before the text block may be deleted:
  * delete the gate from files-actions.ts -> "refuses ... before touching
    storage" goes red (1 failed, 3 passed); reverted
  * delete the gate from actions.ts       -> the uploadAvatar case goes red
    (1 failed, 3 passed); reverted
  * restored -> 4/4 green
Each mutation reddens exactly one case, which is what tells you the two
entrypoints are independently covered rather than one assertion doing both jobs.

Only then was the text block removed, in this same commit, with a note at its
old site saying where the coverage went — so the next reader does not read a
deletion as a loss.

Suite unchanged at 2667 total / 2649 passed / 18 failed (the known macOS pair):
two text scans out, two behavioural cases in. tsc clean, lint 0 errors.
